// pab_ng-main/pab/src/app/shared/model/berichte/ergebnis-b002-uebersicht.ts
export interface ErgebnisB002Uebersicht {
  jahr: number;
  monat: number;
  mitarbeiterPersonalnummer: number;
  mitarbeiterIstAktiv: boolean;
  mitarbeiterVorname: string;
  mitarbeiterNachname: string;
  mitarbeiterKurzname: string;
  sachbearbeiterVorname: string;
  sachbearbeiterNachname: string;
}